feat: Expose onboarding components through an object to not pollute general discoverability
