--- Onboarding/src/main/kotlin/com/infomaniak/core/onboarding/components/DefaultTitleAndDescription.kt
@@ -1,46 +1,0 @@
-/*
- * Infomaniak SwissTransfer - Android
- * Copyright (C) 2025 Infomaniak Network SA
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.infomaniak.core.onboarding.components
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun DefaultTitleAndDescription(
-    title: String,
-    description: String,
-    titleStyle: TextStyle, // TODO: Default to correct value when we have a design system
-    descriptionStyle: TextStyle, // TODO: Default to correct value when we have a design system
-    modifier: Modifier = Modifier.Companion,
-) {
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(16.dp),
-        horizontalAlignment = Alignment.Companion.CenterHorizontally,
-    ) {
-        Text(textAlign = TextAlign.Companion.Center, text = title, style = titleStyle)
-        Text(textAlign = TextAlign.Companion.Center, text = description, style = descriptionStyle)
-    }
-}
